Move ServeStale tests to separate file
We're using the word "failover" to describe falling back through multiple
backends. Serve stale isn't strictly related to this and we now have one
vendor that can do serve stale without yet doing multiple backends.

This is just a straight move. I'll rename and remove checkForSkipFailover()
in subsequent commits for clarity.
